backend/forecast_module/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
certifi==2025.4.26
click==8.2.0
cmdstanpy==1.2.5
colorama==0.4.6
contourpy==1.3.2
cycler==0.12.1
Cython==3.1.0
fastapi==0.115.12
fonttools==4.58.0
h11==0.16.0
holidays==0.73
httptools==0.6.4
idna==3.10
importlib_resources==6.5.2
joblib==1.5.0
kiwisolver==1.4.8
matplotlib==3.10.3
numpy==2.2.5
packaging==25.0
pandas==2.2.3
patsy==1.0.1
pillow==11.2.1
pmdarima==2.0.4
polygon-api-client==1.14.5
prophet==1.1.6
pydantic==2.11.4
pydantic_core==2.33.2
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
PyYAML==6.0.2
scikit-learn==1.6.1
scipy==1.15.3
setuptools==80.4.0
six==1.17.0
sniffio==1.3.1
stanio==0.5.1
starlette==0.46.2
statsmodels==0.14.4
threadpoolctl==3.6.0
tqdm==4.67.1
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
uvicorn==0.34.2
watchfiles==1.0.5
websockets==14.2
